Add files via upload
Q1. Calculating the inverse of a 2x2 matrix without using numpy
Code a function to calculate the determinant of 2x2 matrix

Code a function that interchange the diagonal elements of a 2x2 matrix and inverse the sign of the off diagonal elements.

Code a function to compute the inverse of 2x2 matrix based on the two previous functions if it exists

Q2. Coding a Python code to inverse a 3x3 matrix in order to solve a linear system (no numpy.linalg.inv allowed) with 3 constraints and 3 variables:
Coding a function that checks if a 3x3 matrix is invertible
Coding a function that generates the transpose of a 3x3 matrix
Coding a function that generates the matrix of minors of a 3x3 matrix
Coding a function that generates the matrix of cofactors of a 3x3 matrix
Coding a function that generates the inverse of a 3x3 matrix if it exists